project renamed culturo && code cleaned up && Swagger implemented

+++ api-culturo/src/users/users.service.ts
@@ -7,12 +7,13 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Role } from 'src/entities/role.entity';
 import { Utilisateur } from 'src/entities/utilisateur.entity';
 import { Repository } from 'typeorm';
-import { RegisterDTO } from './dtos/register.dto';
+import { RegisterDTO } from './dtos/register.user.dto';
 import bcrypt from 'node_modules/bcryptjs';
 import { LoginDTO } from './dtos/login.dto';
 import { JwtService } from '@nestjs/jwt';
 import { AccessTokenType, JWTPayloadType } from 'src/utils/types';
 import { ConfigService } from '@nestjs/config';
+import { UpdateUserDTO } from './dtos/update.user.dto';
 
 @Injectable()
 export class UsersService {
@@ -24,35 +25,76 @@ export class UsersService {
     private readonly jwtService: JwtService,
   ) {}
 
-  // Récupérer tous les utilisateurs
+  /**
+   *
+   * @returns
+   */
   async getAllUsers(): Promise<Utilisateur[]> {
     return await this.userRepository.find();
   }
 
-  // Récupérer un utilisateur par id
+  /**
+   *
+   * @param id
+   * @returns
+   */
   async getUserById(id: number): Promise<Utilisateur | null> {
     return this.userRepository.findOneBy({ id_utilisateur: id });
   }
 
-  // Supprimer un utilisateur
+  /**
+   *
+   * @param id
+   * @returns
+   */
   async delUser(id: number): Promise<{ msg: string }> {
+    const user = await this.userRepository.findOne({
+      where: { id_utilisateur: id },
+    });
+
+    if (!user) {
+      throw new NotFoundException('Utilisateur introuvable');
+    }
     await this.userRepository.delete({ id_utilisateur: id });
     return { msg: 'Utilisateur supprimé avec succès' };
   }
 
-  // Mettre à jour un utilisateur
+  /**
+   *
+   * @param updateData
+   * @returns
+   */
   async updateUser(
-    user: Utilisateur,
+    updateData: UpdateUserDTO,
   ): Promise<{ status: number; msg: string }> {
-    console.log(user);
-    try {
-      await this.userRepository.update(user.id_utilisateur, user);
-      await this.userRepository.save(user);
-      return { status: 200, msg: 'Update OK' };
-    } catch (error) {
-      return { status: 500, msg: 'Update n’a pas marché' };
+    const user = await this.userRepository.findOne({
+      where: { id_utilisateur: updateData.id_utilisateur },
+      relations: ['role'],
+    });
+    if (!user) {
+      // throw new NotFoundException('Utilisateur introuvable');
+      return { status: 404, msg: 'Not Found User ' };
     }
+    // Mise à jour du rôle si fourni
+    if (updateData.id_role) {
+      const role = await this.roleRepository.findOne({
+        where: { id_role: updateData.id_role },
+      });
+      if (!role) {
+        throw new BadRequestException('Role introuvable');
+      }
+      user.role = role;
+    }
+    // Mise à jour du mot de passe si changement
+    if (updateData.hpassword) {
+      user.hpassword = await bcrypt.hash(updateData.hpassword, 10);
+    }
+    // Mise à jour des autres champs
+    Object.assign(user, updateData);
+    await this.userRepository.save(user);
+    return { status: 200, msg: 'Utilisateur mis à jour avec succès' };
   }
+
   /**
    * Methode of create user
    * @param registerDto data for creating user
@@ -83,7 +125,7 @@ export class UsersService {
     };
 
     const accessToken = await this.jwtService.signAsync(payload);
-    return { id: savedUser.id_utilisateur ,accessToken };
+    return { id: savedUser.id_utilisateur, accessToken };
   }
 
   /**
@@ -104,7 +146,7 @@ export class UsersService {
       id_role: userFromDb.id_role,
     };
     const accessToken = await this.jwtService.signAsync(payload);
-    return { id: userFromDb.id_utilisateur , accessToken };
+    return { id: userFromDb.id_utilisateur, accessToken };
   }
 
   /**
@@ -112,7 +154,7 @@ export class UsersService {
    * @param id  Id of the user logged
    * @returns the user from the database
    */
-  public async getCurrentUser(id : number) {
+  public async getCurrentUser(id: number) {
     const userFromDb = await this.userRepository.findOne({
       where: { id_utilisateur: id },
     });
@@ -120,4 +162,3 @@ export class UsersService {
     return userFromDb;
   }
 }
-
